refactoring of sample generators and builders
