+++ backend/api/SampleData/Builders/DrainageStrategyBuilder.cs
@@ -1,6 +1,6 @@
 using api.Models;
 
-namespace api.SampleData;
+namespace api.SampleData.Builders;
 
 public class DrainageStrategyBuilder : DrainageStrategy
 {
