+++ backend/api/SampleData/Builders/TopsideBuilder.cs
@@ -1,6 +1,6 @@
 using api.Models;
 
-namespace api.SampleData;
+namespace api.SampleData.Builders;
 
 public class TopsideBuilder : Topside
 {
