+++ backend/api/SampleData/Generators/SampleAssetGenerator.cs
@@ -1,6 +1,7 @@
 using api.Models;
+using api.SampleData.Builders;
 
-namespace api.SampleData;
+namespace api.SampleData.Generators;
 
 public static class SampleAssetGenerator
 {
